GH-2 - Enhance API.
This change exposes a new API via two annotations: `@NeedsCausalCluster` still is used to indicate the need for a causal cluster during tests while `@CausalCluster` replaces `@Neo4jUri`. `@CausalCluster` can now be used to inject one or all URIs of the Causal Cluster into one String or URI or into collection fields of that type. It also can be used to inject the available cluster instance as `Neo4jCluster` into a field of that type. The `Neo4jCluster` than provides access to all `Neo4jServer` of that cluster.

Co-authored-by: Michael Simons <michael@simons.ac>