GH-25 - Make the number of read replicas configurable.
We already had an attribute to configure the number of read replicas, but didn’t expose or use it. This change exposes the attribute on the `@NeedsCausalCluster` annotation and uses it accordingly.

The Neo4jServer now has an attribute indicating it’s type. The entry points to the cluster (both the random single one and the list of all Uris) will include only core servers. For explicitly working with read replicas, the list of all servers needs to be retrieved and filtered as needed. 

This closes #22.